refactored tests to have pqc tests under ShowcaseTest.php
