+++ .github/workflows/conformance-tests-gax-showcase.yaml
@@ -16,10 +16,10 @@ jobs:
         name: GAPIC Showcase Conformance Tests
         steps:
         - name: Checkout code
-          uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 #v7.0.1
+          uses: actions/checkout@v7
 
         - name: Setup PHP
-          uses: shivammathur/setup-php@f3e473d116dcccaddc5834248c87452386958240 #2.37.2
+          uses: shivammathur/setup-php@v2
           with:
             php-version: '8.2'
             extensions: grpc-1.83.0
